+++ Tests/StencilTests/IfNodeSpec.swift
@@ -27,6 +27,28 @@ func testIfNode() {
         try expect(falseNode?.text) == "false"
       }
 
+      $0.it("can parse an if with complex expression") {
+        let tokens: [Token] = [
+          .block(value: "if value == \"test\" and not name"),
+          .text(value: "true"),
+          .block(value: "else"),
+          .text(value: "false"),
+          .block(value: "endif")
+        ]
+
+        let parser = TokenParser(tokens: tokens, namespace: Namespace())
+        let nodes = try parser.parse()
+        let node = nodes.first as? IfNode
+        let trueNode = node?.trueNodes.first as? TextNode
+        let falseNode = node?.falseNodes.first as? TextNode
+
+        try expect(nodes.count) == 1
+        try expect(node?.trueNodes.count) == 1
+        try expect(trueNode?.text) == "true"
+        try expect(node?.falseNodes.count) == 1
+        try expect(falseNode?.text) == "false"
+      }
+
       $0.it("can parse an ifnot block") {
         let tokens: [Token] = [
           .block(value: "ifnot value"),